+++ src/hooks/useMediaQuery/useMediaQuery.ts
@@ -1,17 +1,25 @@
-import React from "react";
+import { useLayoutEffect, useState } from "react";
+import { getWindow } from "../../__internal__/dom/globals";
 
-export default function useMediaQuery(queryInput: string): boolean {
+export default function useMediaQuery(queryInput: string): boolean | undefined {
   const query = queryInput.replace(/^@media( ?)/m, "");
 
-  const [match, setMatch] = React.useState(() => false);
+  const [match, setMatch] = useState<boolean | undefined>(undefined);
 
-  React.useEffect(() => {
-    const queryList = window.matchMedia(query);
-    const updateMatch = () => {
-      setMatch(queryList.matches);
-    };
+  useLayoutEffect(() => {
+    const browserWindow = getWindow();
+
+    /* istanbul ignore if */
+    if (!browserWindow) {
+      return undefined;
+    }
+
+    const queryList = browserWindow.matchMedia(query);
+    const updateMatch = () => setMatch(queryList.matches);
+
     updateMatch();
     queryList.addEventListener("change", updateMatch);
+
     return () => {
       queryList.removeEventListener("change", updateMatch);
     };