+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+## [153.5.2](https://github.com/Sage/carbon/compare/v153.5.1...v153.5.2) (2025-04-23)
+
+### Bug Fixes
+
+* ensure there is no flash of reduce motion text in Loader components when page is reloaded ([c65e0b2](https://github.com/Sage/carbon/commit/c65e0b20c259973eb11c5e09e3ae460b5cc823a1)), closes [#7240](https://github.com/Sage/carbon/issues/7240)
+
+## [153.5.1](https://github.com/Sage/carbon/compare/v153.5.0...v153.5.1) (2025-04-23)
+
+### Bug Fixes
+
+* **text-editor, note:** ensure links have correct styles ([ac9e0b8](https://github.com/Sage/carbon/commit/ac9e0b899f40366d25b28621c10d528efbf639e4))
+
+## [153.5.0](https://github.com/Sage/carbon/compare/v153.4.0...v153.5.0) (2025-04-23)
+
+### Features
+
+* **fieldset, label:** add de-DE, es-ES, fr-CA and fr-FR support for optional indicators ([e78e212](https://github.com/Sage/carbon/commit/e78e212463177348e23a0961264d54bf10390cc8))
+
 ## [153.4.0](https://github.com/Sage/carbon/compare/v153.3.1...v153.4.0) (2025-04-16)
 
 ### Features